requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-community==0.0.4
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
twilio==8.10.0
python-dotenv==1.0.0
mangum==0.17.0
